share NewtypeSpec between versions
