+++ core/shared/src/test/scala-2.x/zio/prelude/NewtypeSpecVersionSpecific.scala
@@ -1,0 +1,13 @@
+package zio.prelude
+
+object NewtypeSpecTypes {
+  type Natural = Natural.Type
+  object Natural extends Subtype[Int] {
+    def refinement =
+      refine(Refinement.greaterThanOrEqualTo(0))
+
+    val two: Natural = Natural(2)
+
+    def unsafeWrap(int: Int): Natural = wrap(int)
+  }
+}
